+++ kegal/__init__.py
@@ -48,7 +48,7 @@ from .validators import (
 from . import llm
 from .llm.llm_model import LLMTool
 
-__version__ = "0.1.4.1"
+__version__ = "0.1.4.5"
 
 __all__ = [
     # Graph models